feat: algin time with threshold
Previously, a ground truth was found for *every* estimated pose
by taking the one that is closest in time, even if that might be
long after the estimate.
With the threshold, it is also possible to compute how much of
the ground truth is actually used for evaluation and where (due
to a lower publishing rate or due to temporary loss of tracking)
orb slam does not give any estimates. Without this, high scores
of the pipeline could be achieved simply by making estimates only
in the easy portions of any sequence.