fix: relative error in evaluate_trajectory
The previous implementation scaled to have the total lengths of
sub-trajectories be the same. Observing this on data, one can
observe that when the estimated trajectory is more noisy than the
gt, this technique leads to excessive shrinkage of the estimated
trajectory.
Improved version: compute sub-trajectory lengths by considering
the difference between the first and the last position in the
sub-trajectory.
Also: remove the scale-drift graphic in visualisation for stereo.